Add concept of turn in combat
Right now there are multiple frames in which pressing G will end the
players turn because there is no sleeping during an AI turn. This
causes the turns to rapidly alternate during the few frames the G
key is pressed. It will be fixed as soon as AI turns are actually
implemented.