Add Woodsman and add Brambles skill
Noticed that Dog's movement after HitandRun can take a LONG time to
display - I'll need to debug where the problem is (probably in
"GenerateTileCircle") and optimize. Since the Woodsman uses the whole
grid for Brambles, and the lag isn't apparent there, it seemed
surprising. However the Brambles overrides the recursion, so it must be
that.