Fix melee AI bug
Entities would get stuck in a tile diagonal to the target, or constantly
"strafe" as it attacked the entity. Adding the tile it currently was
standing on in the score computation, as well as choosing the minimum
distance to the nearest "attackable" tile solved the problem... I think.

It's hard to repro AI bugs because it's difficult to "replay" combat - a
tool to replay steps in combat would be killer for debugging AI issues.
Should seriously consider writing a playback tool if I want to continue
after the prototype.