Create gauntlet scene
This breaks some things for the main scene, so I'm creating it on a
separate branch.

I'll need to refactor some stuff if I want to make sure that it all
works cleanly. Right now this commit would have some bugs if I didn't
comment out some of the dialog stuff.